Makes 4-mer histograms
This is the code that reads in an array of z-scores or tuds and
associated pals, then outputs folder with a series of png files that
contain one histogram per pal, named pal.png. This script is hard-coded
to work when the # of palindromes = 16, i.e. when dealing with pals in
4mers.