Working OOP infrastructure
This program script is designed to run from the command line. It takes
in one argument (the name of the fasta), and creates a list of pairwise
objects from each group of 4 lines in the fasta. This script also has a
printing method, and the only thing the main method here does is print
out the list of pairwise objects. I’m pushing it because all the syntax
to make the objects work is functional, and tunable from the command
line.