Like Pairwise1 but better
This script actually accesses the entire folder of  phaedrus pairwise
alignments, and writes the output to a file called pairwise_output. The
only thing it does in this script is print out the name of each
alignment from the gpX file “query 1— whatever” then \n then same thing
for gpX+1 etc until it reaches the end of the gp directory. The reason
I’m pushing it is because the flow of control works for file I/O, and
if I break over the course of the next few min as I try to generate
more meaningful output files, I want to have this as a backstop.